refactor(frontend): Calendar 컴포넌트에서 util 함수를 별도 파일로 분리
